fix overflow in constant conversion

+++ janus-uae/src/audio.cpp
@@ -1633,7 +1633,7 @@ void audio_reset (void)
 		for (i = 0; i < AUDIO_CHANNELS_MAX; i++) {
 			cdp = &audio_channel[i];
 			memset (cdp, 0, sizeof *audio_channel);
-			cdp->per = PERIOD_MAX - 1;
+			cdp->per = (int)PERIOD_MAX - 1;
 			cdp->vol = 0;
 			cdp->evtime = MAX_EV;
 		}
